* server-src/amcheck.c, server-src/driver.c,   server-src/planner.c: Cast pid_t to int before print with %d
git-svn-id: https://svn.code.sf.net/p/amanda/code/amanda/trunk@7431 a8d146d6-cc15-0410-8900-af154a0219e0

+++ server-src/amcheck.c
@@ -2750,7 +2750,7 @@ handle_result(
        them explicitly */
     {	int pid;
 	while((pid = waitpid(-1, NULL, WNOHANG))> 0) {
-	    g_debug("amcheck reap: %d", pid);
+	    g_debug("amcheck reap: %d", (int)pid);
 	}
     }
 }
